Phase 3: Advanced Query Features — OR, Navigation, Statistics (#141) (#158)

+++ test/Typhon.Engine.Tests/Data/Query/ViewMultiComponentTests.cs
@@ -15,8 +15,8 @@ class ViewMultiComponentTests : TestBase<ViewMultiComponentTests>
         KeyType keyType, CompareOp compareOp, long threshold, byte componentTag = 0) =>
         new()
         {
-            FieldIndex = fieldIndex,
-            FieldOffset = fieldOffset,
+            FieldIndex = (byte)fieldIndex,
+            FieldOffset = (ushort)fieldOffset,
             FieldSize = fieldSize,
             KeyType = keyType,
             CompareOp = compareOp,
